Fix header style and quicklinks bugs
