+++ src/pages/home/components/QuickLinks/styles.scss
@@ -1,4 +1,9 @@
 .button {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  text-align: center;
+
   &:hover,
   &:active,
   &:focus {
